Add request migration and refactor models
