+++ .env.example
@@ -1,1 +1,2 @@
-FLASK_APP=run
+FLASK_APP=main
+FLASK_ENV=development
