+++ app/__init__.py
@@ -1,10 +1,8 @@
 from flask import Flask
-from config import Config
 from flask_sqlalchemy import SQLAlchemy
+from config import Config
 
-app = Flask(__name__)
-app.config.from_object(Config)
-db = SQLAlchemy(app)
+db = SQLAlchemy()
 
 
 def create_app():
