--- app/models/user.py
@@ -1,20 +1,0 @@
-from app import db
-
-
-class User(db.Model):
-    __tablename__ = 'users'
-
-    id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80), unique=False, nullable=False)
-    email = db.Column(db.String(120), unique=True, nullable=False)
-
-    role_id = db.Column(db.Integer, db.ForeignKey('roles.id'), nullable=False)
-    role = db.relationship('Role', backref=db.backref('users', lazy=True))
-
-    def __init__(self, name, email, role):
-        self.name = name
-        self.email = email
-        self.role = role
-
-    def __repr__(self):
-        return f'<User {self.name}>'
